Rename flat/sharedexception.h -> flat/exceptionobj.h
